+++ client/src/components/screens/post/PostDetailScreen.js
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from "react";
-import { PostContext } from "../../post";
-import { GlobalStoreContext } from "../../store";
+import { PostContext } from "../../../post";
+import { GlobalStoreContext } from "../../../store";
 import { useNavigate } from "react-router-dom";
 import {
   Box,
@@ -14,7 +14,7 @@ import {
 import { Typography } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import AddIcon from "@mui/icons-material/Add";
-import {PostComment} from "../index"
+import {PostComment} from "../../index"
 
 export default function PostDetailScreen() {
   const navigate = useNavigate();
